Change ball.png to a "real" ball image.
The png was generated from an svg using inkscape.

The svg came from here:

https://openclipart.org/download/65035/1275981996.svg
https://openclipart.org/detail/65035/beach-ball
https://openclipart.org/user-detail/laobc

Its licence is:

https://openclipart.org/share

Unlimited Commercial Use

We try to make it clear that you may use all clipart from Openclipart
even for unlimited commercial use. We believe that giving away our
images is a great way to share with the world our talents and that
will come back around in a better form.

May I Use Openclipart for?

We put together a small chart of as many possibilities and questions
we have heard from people asking how they may use Openclipart. If you
have an additional question, please email love@openclipart.org.

All Clipart are Released into the Public Domain.

Each artist at Openclipart releases all rights to the images they
share at Openclipart. The reason is so that there is no friction in
using and sharing images authors make available at this website so
that each artist might also receive the same benefit in using other
artists clipart totally for any possible reason.

Need a Legal Disclaimer?

We use the Creative Commons Zero 1.0 Public Domain License every time
an artist uploads a piece of clipart to Openclipart to make it clear
the artist is releasing the creative work for anyone to use for any
reason, even commercially. This act of "sharing" is the foundation
Openclipart is based upon. Sometimes people want a disclaimer or legal
assurance that our clipart is safe to be used. This extra security
request often comes from people working for companies of all sizes
and, or, publishers. If you need this type of document, please visit,
save, download and, or print the following for your needs: Creative
Commons Zero 1.0 License.